Excluding com.jcraft.jsch.agentproxy from exported packages
Also ensuring it's no longer transitively included in the JAR.
This means OSGi users will also need to import the
jsch.agentproxy.core bundle, which *should* be providing the core
files.

That, in turn, avoids a "uses constraint violation" when trying to
load both the usocket-nc and usocket-jna bundles if the core bundle
is not being imported (which is currently not necessary, since
connector-factory exports the core package too).